Remote all preset values from the run-time configs of the Gnosis networks
The Ethereum specs define which values belong to a const preset (and are
suitable for specifying at compile-time in client builds) and which are
expected to be configurable at run-time.

The removed values here are part of the const preset set and a fully
spec-compliant client may produce a warning or error when encountering
them in the run-time config file.